refac (Price) : BigDecimal 매개변수 정적 팩터리 메소드 삭제 리팩터링

+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -76,6 +76,7 @@ public class MenuProduct {
 	public Price getTotalPrice() {
 		return Price.from(product
 			.getPrice()
-			.multiply(BigDecimal.valueOf(quantity)));
+			.multiply(BigDecimal.valueOf(quantity))
+			.intValue());
 	}
 }
